+++ .github/workflows/ci.yml
@@ -8,11 +8,12 @@ jobs:
     timeout-minutes: 20
     steps:
       - uses: actions/checkout@v2
-      - uses: actions/setup-node@v1
+      - uses: actions/setup-node@v2
         with:
-          node-version: 14
+          node-version: 12
+          cache: 'npm'
       - name: Install
-        run: npm install --prefer-offline --frozen-lockfile --no-progress --non-interactive
+        run: npm install
       - name: Format check
         run: npm run format:check
       - name: Lint
