+++ .nvmrc
@@ -1,1 +1,1 @@
-14
+12
