+++ jest.config.js
@@ -4,7 +4,7 @@ module.exports = {
   testEnvironment: 'node',
   globals: {
     'ts-jest': {
-      tsConfig: '<rootDir>/test/tsconfig.json',
+      tsconfig: '<rootDir>/test/tsconfig.json',
     },
   },
   testPathIgnorePatterns: ['<rootDir>/build/', '<rootDir>/src/'],
