+++ src/version-api/version.api.ts
@@ -1,11 +1,11 @@
 import globalAxios, { AxiosInstance, AxiosPromise, AxiosRequestConfig } from 'axios';
 
-import { RequestArgs } from '../api';
+import { RequestArgs } from '../tonic';
 import { Configuration } from '../configuration';
 import { getRequestOptionsWithApiKey } from '../lib/get-request-options-with-api-key';
 import { VersionAPIAxiosParamCreator, VersionAPI } from './version.api.types';
 
-const versionApiAxiosParamCreator = (configuration: Configuration): VersionAPIAxiosParamCreator => {
+const versionApiRequestArgsCreator = (configuration: Configuration): VersionAPIAxiosParamCreator => {
   return {
     getVersionRequestArgs: async (options: AxiosRequestConfig = {}): Promise<RequestArgs> => {
       const requestPath = `/api/Version`;
@@ -19,12 +19,12 @@ const versionApiAxiosParamCreator = (configuration: Configuration): VersionAPIAx
   };
 };
 
-export const VersionApi = (configuration: Configuration): VersionAPI => {
+const versionApi = (configuration: Configuration): VersionAPI => {
   return {
     async getVersion(
       options?: AxiosRequestConfig
     ): Promise<(axios?: AxiosInstance, basePath?: string) => AxiosPromise<number>> {
-      const requestArgs = await versionApiAxiosParamCreator(configuration).getVersionRequestArgs(options);
+      const requestArgs = await versionApiRequestArgsCreator(configuration).getVersionRequestArgs(options);
 
       return (axios: AxiosInstance = globalAxios, basePath): AxiosPromise<number> => {
         const axiosRequestArgs = { ...requestArgs.options, url: basePath + requestArgs.url };
@@ -34,3 +34,5 @@ export const VersionApi = (configuration: Configuration): VersionAPI => {
     },
   };
 };
+
+export { versionApi };
