+++ tsconfig.json
@@ -5,7 +5,8 @@
     "esModuleInterop": true,
     "skipLibCheck": true,
     "allowSyntheticDefaultImports": true,
-    "target": "es6",
+    "target": "es2019",
+    "lib": ["es2019", "es2020.bigint", "es2020.string", "es2020.symbol.wellknown"],
     "moduleResolution": "node",
     "sourceMap": true,
     "declaration": true,
